+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Devop
 
+- Fixes gasPrice when generating offline. []()
 - Update u2f-api [#2135](https://github.com/kvhnuke/etherwallet/pull/2135)
 
 ### Release v3.24.0